Renamed folder "server/db-models" to "server/data-source-models", in order to make it more generic, and moved file "pagination-output.ts" to this new folder (also adding the suffix "-model" to this file)
