+++ src/server/model-converters/user-model-converter.ts
@@ -1,6 +1,6 @@
-import { UserCreationModel, UserModel, UserUpdateModel } from '../db-models/user-models';
+import { UserCreationModel, UserModel, UserUpdateModel } from '../data-source-models/user-models';
 import { User, UserCreationInput, UserPaginationResult, UserUpdateInput } from '../graphql/types';
-import { PaginationOutputModel } from '../types/pagination-output';
+import { PaginationOutputModel } from '../data-source-models/pagination-output-model';
 
 // This class is meant to decouple the models from graphQL API from the models that are used with our database tables.
 // With UserModelConverter, we stopped using graphQL input models directly when creating or updating a user, for example.
